+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.9 branch
+### 2.9.4
+* updated for MeCms 2.18.0.
+
 ### 2.9.3
 * widgets (cells) use collections;
 * tests that require a network connection have been marked with the